requirements.txt
torch
gym
pyyaml
tensorboard
pytest
